Remove reudndant functions from string column, moving the ones that should be generic into ColumnBase.
